feature: AppTrainingPage component, improved UseOfEnglish component
