+++ app/src/views/UseOfEnglish.vue
@@ -1,52 +1,33 @@
 <template>
-    <main v-if="questions.length">
-        <teleport to="body">
-            <app-modal v-if="showInstruction" title="Инструкция" @close="showInstruction = false">
-                <div>
-                    <p>
-                        Преобразуйте слова, напечатанные заглавными буквами так, чтобы они грамматически и лексически
-                        соответствовали содержанию текстов.
-                    </p>
-                    <p>
-                        Заполните пропуски полученными словами.
-                    </p>
-                    <p>
-                        Слова вводите заглавными буквами, без пробелов, как в экзаменационном бланке.
-                    </p>
-                    <p>
-                        Глагольные формы вводите без сокращений.
-                    </p>
-                    <button class="btn btn-block btn-centered primary" @click="showInstruction = false">ОК</button>
-                </div>
-            </app-modal>
-        </teleport>
+    <main>
+        <AppTrainingPage :topic="topic" ref="page" @check-answers="checkAnswers">
+            <template #training-instruction>
+                <p>
+                    Преобразуйте слова, напечатанные заглавными буквами так, чтобы они грамматически и лексически
+                    соответствовали содержанию текстов.
+                </p>
+                <p>
+                    Заполните пропуски полученными словами.
+                </p>
+                <p>
+                    Слова вводите заглавными буквами, без пробелов, как в экзаменационном бланке.
+                </p>
+                <p>
+                    Глагольные формы вводите без сокращений.
+                </p>
+            </template>
 
-        <AppTrainingHeader :topic="topic" @show-instruction="showInstruction = true"/>
-
-        <UseOfEnglishCard
-                v-for="(question, i) in questions"
-                :key="question._id"
-                :id="question._id"
-                :question="question.task"
-                :origin="question.origin"
-                :ref="el => { if (el) uoeCards[i] = el }"
-        />
-        <div class="buttons-group">
-            <button class="btn primary send-answers-btn" :disabled="isChecking"
-                    v-if="!isChecked" @click="checkAnswers">
-                Проверить
-            </button>
-            <button class="btn secondary" @click="$router.go(-1)">Выход</button>
-        </div>
-
-        <teleport to="body">
-            <app-modal v-if="showResult" :title="result" @close="showResult = false">
-                <div>
-                    <p>Вы можете посмотреть свои ошибки и правильные ответы</p>
-                    <button class="btn btn-block btn-centered primary" @click="showResult = false">ОК</button>
-                </div>
-            </app-modal>
-        </teleport>
+            <template #training-content>
+                <UseOfEnglishCard
+                        v-for="(question, i) in questions"
+                        :key="question._id"
+                        :id="question._id"
+                        :question="question.task"
+                        :origin="question.origin"
+                        :ref="el => { if (el) uoeCards[i] = el }"
+                />
+            </template>
+        </AppTrainingPage>
     </main>
 </template>
 
@@ -55,36 +36,32 @@ import UseOfEnglishCard from '@/components/cards/UseOfEnglishCard';
 import {computed, onBeforeUpdate, onMounted, ref} from 'vue';
 import {API} from '@/services/api';
 import {replaceCharSequence} from '@/utils/replaceCharSequence';
-import AppModal from '@/components/AppModal';
-import AppTrainingHeader from '@/components/AppTrainingHeader';
+import AppTrainingPage from '@/components/AppTrainingPage';
 
 export default {
     name: 'UseOfEnglish',
-    components: {AppTrainingHeader, UseOfEnglishCard, AppModal},
+    components: {AppTrainingPage, UseOfEnglishCard},
     props: {
         topic: {
             type: String
         }
     },
     setup(props) {
+        const page = ref(null);
         const questions = ref([]);
         const uoeCards = ref([]);
-        const showInstruction = ref(false);
 
         onMounted(async () => {
             API.getUoeTraining(props.topic)
                 .then((res) => {
                     questions.value = res.data.questions;
-                    showInstruction.value = true;
+                    page.value.setShowInstruction(true);
                 })
                 .catch((err) => console.log(err));
         });
 
         onBeforeUpdate(() => uoeCards.value = []);
 
-        const isChecking = ref(false);
-        const isChecked = ref(false);
-        const showResult = ref(false);
         const rightAnswers = ref(null);
         const result = computed(() => {
             const resultRatio = (rightAnswers.value === null || questions.value === null) ? null :
@@ -93,7 +70,7 @@ export default {
         });
 
         const checkAnswers = async () => {
-            isChecking.value = true;
+            page.value.setIsChecking(true);
 
             const userAnswers = uoeCards.value.map((cardComponent) => cardComponent.getAnswerData());
             const checkResultResponse = await API.checkTraining('use-of-english', userAnswers);
@@ -109,19 +86,14 @@ export default {
                     rightAnswer, {tagWrapper: 'strong'});
             });
 
-            isChecking.value = false;
-            isChecked.value = true;
-            showResult.value = true;
+            page.value.setResult(result.value);
         };
 
         return {
+            page,
             questions,
             uoeCards,
-            isChecking,
-            isChecked,
             result,
-            showResult,
-            showInstruction,
             checkAnswers
         };
     }
@@ -134,21 +106,5 @@ export default {
 main {
   width: 50%;
   margin: 32px auto;
-}
-
-.buttons-group {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 16px;
-
-  button {
-    width: 100%;
-    height: 50px;
-    border-radius: 8px;
-    font-size: 18px;
-    font-family: Inter, Roboto, Oxygen, Fira Sans, Helvetica Neue, sans-serif;
-  }
 }
 </style>
